==> credit evaluation_univariate/code_2nd week/02. CB,SP 0등급 및 기부도 대상 선정.ipynb <==
import pandas as pd
import numpy as np
import os
import warnings
warnings.filterwarnings('ignore')
from datetime import datetime

# <font color='orange'>01. load data

rootpath = r'C:\Users\w10\Desktop\신용평가모형 세미나\2주차\데이터\temp'

os.listdir(rootpath)

total_df = pd.read_pickle(os.path.join(rootpath, 'total_df.pkl'), )

inner_total_df = pd.read_csv(os.path.join(rootpath, 'temp1_inner_total_df.csv'))

inner_total_df.head(2)

# <font color='orange'>02. CB, SP 0등급 대상 선정

# 등급별 현황파악
print('CB등급별 건수')
print(total_df['Grd_RK0400_000'].value_counts().sort_index(),)
print('SP등급별 건수')
print(total_df['Grd_RK0400_700'].value_counts().sort_index())

# CB, SP 0등급 적중 함수 생성
def grade_zero(cb, sp):
    result = 0
    if cb==0 or sp==0:
        result = 1
        
    return result

# input columns list 생성
input_cols = ['Grd_RK0400_000', 'Grd_RK0400_700']

# CB, SP 0등급 항목 생성
total_df['GRADE_ZERO_FG'] = total_df[input_cols].apply(lambda x: grade_zero(x['Grd_RK0400_000'], x['Grd_RK0400_700']), axis = 1)

total_df['GRADE_ZERO_FG'].sum()

# 내부항목 dataframe 적용
inner_total_df['GRADE_ZERO_FG'] = total_df['GRADE_ZERO_FG']

inner_total_df['GRADE_ZERO_FG'].sum()

# <font color='orange'>03. 기부도 대상 선정

# def _pre_bad(BS0000145, BS0000172, BS0000169,
#             BS0000171, B12000200, B22000200,
#             PS0000296_1, P27000100):
    
#     BS0000145_FG = 0
#     BS0000172_FG = 0 
#     BS0000169_FG = 0
#     BS0000171_FG = 0
#     B12000200_FG = 0
#     B22000200_FG = 0
#     PS0000296_1_FG = 0
#     P27000100_FG = 0
    
#     if BS0000145 > 0:
#         BS0000145_FG = 1
        
#     if BS0000172 > 0:
#         BS0000172_FG = 1
        
#     if BS0000169 > 0:
#         BS0000169_FG = 1
        
#     if BS0000171 > 0:
#         BS0000171_FG = 1
        
#     if B12000200 > 0:
#         B12000200_FG = 1
        
#     if B22000200 > 0:
#         B22000200_FG = 1
        
#     if PS0000296_1 > 0:
#         PS0000296_1_FG = 1
        
#     if P27000100 > 0:
#         P27000100_FG = 1
        
#     return BS0000145_FG,BS0000172_FG, BS0000169_FG,BS0000171_FG,B12000200_FG,B22000200_FG,PS0000296_1_FG,P27000100_FG,
    
    
# total_df[result_cols] = total_df[need_cols].apply(lambda x: _pre_bad(x['BS0000145'],
#                                                                     x['BS0000172'],
#                                                                     x['BS0000169'],
#                                                                     x['BS0000171'],
#                                                                     x['B12000200'],
#                                                                     x['B22000200'],
#                                                                     x['PS0000296_1'],
#                                                                     x['P27000100'],), axis = 1, result_type = 'expand')

def pre_bad(col):
    result = 0
    if col>0:
        result=1
        
    return result

# output 항목명
result_cols = [ 'PRE_BAD_FG_1',
               'PRE_BAD_FG_2',
               'PRE_BAD_FG_3',
               'PRE_BAD_FG_4',
               'PRE_BAD_FG_5',
               'PRE_BAD_FG_6',
               'PRE_BAD_FG_7',
               'PRE_BAD_FG_8',]

# input 항목명
need_cols = [ 'BS0000145',
            'BS0000172', 
            'BS0000169',
            'BS0000171',
            'B12000200',
            'B22000200',
            'PS0000296_1',
            'P27000100',]

for in_col, out_col in zip(need_cols, result_cols):
    total_df[out_col] = total_df[in_col].apply(lambda x: pre_bad(x))

total_df[result_cols].sum()

inner_total_df[result_cols] = total_df[result_cols]

### 최종 기부도 대상 항목 생성

# def _total_pre_bad(BS0000145_FG,BS0000172_FG, BS0000169_FG,BS0000171_FG,B12000200_FG,
#                    B22000200_FG,PS0000296_1_FG,P27000100_FG,):
#     total_pre_bad = 0
#     if (BS0000145_FG > 0 or BS0000172_FG > 0 or  BS0000169_FG > 0 or BS0000171_FG > 0 or B12000200_FG > 0 or
#                    B22000200_FG > 0 or PS0000296_1_FG > 0 or P27000100_FG > 0):
#         total_pre_bad = 1
        
#     return total_pre_bad

# total_df['TOTAL_PRE_BAD_FG'] = total_df[need_cols].apply(lambda x: _total_pre_bad(x['PRE_BAD_FG_1'],
#                                                                     x['PRE_BAD_FG_2'],
#                                                                     x['PRE_BAD_FG_3'],
#                                                                     x['PRE_BAD_FG_4'],
#                                                                     x['PRE_BAD_FG_5'],
#                                                                     x['PRE_BAD_FG_6'],
#                                                                     x['PRE_BAD_FG_7'],
#                                                                     x['PRE_BAD_FG_8'],), axis = 1,)

need_cols = [ 'PRE_BAD_FG_1',
               'PRE_BAD_FG_2',
               'PRE_BAD_FG_3',
               'PRE_BAD_FG_4',
               'PRE_BAD_FG_5',
               'PRE_BAD_FG_6',
               'PRE_BAD_FG_7',
               'PRE_BAD_FG_8',]

# 위에서 생성한 기부도 요건들을 or조건 적용
total_df['TOTAL_PRE_BAD_FG'] = total_df[need_cols].sum(axis=1)
total_df['TOTAL_PRE_BAD_FG'] = total_df['TOTAL_PRE_BAD_FG'].apply(lambda x: 1 if x>0 else 0)

inner_total_df['TOTAL_PRE_BAD_FG'] = total_df['TOTAL_PRE_BAD_FG']

inner_total_df.head(2)

# <font color='orange'>04. 저장

save_path = r'C:\Users\w10\Desktop\신용평가모형 세미나\2주차\데이터\temp'

inner_total_df.to_csv(os.path.join(save_path, 'temp2_inner_total_df.csv'), index = False)
