==> credit evaluation_univariate/code_2nd week/04. SEG 선정.ipynb <==
import pandas as pd
import numpy as np
import os
import warnings
warnings.filterwarnings('ignore')
from datetime import datetime

# <font color='orange'>01. load data 

rootpath = r'C:\Users\w10\Desktop\신용평가모형 세미나\2주차\데이터\temp'

os.listdir(rootpath)

total_df = pd.read_pickle(os.path.join(rootpath, 'total_df.pkl'), )

inner_total_df = pd.read_csv(os.path.join(rootpath, 'temp3_inner_total_df.csv'), )

total_df.head(2)

inner_total_df.head(2)

# <font color='orange'>02. 연소득 컬럼 생성

* A5WC0000008000 웰컴크레디라인대부최종연소득
* A5WC0000000200 웰컴크레디라인대부최종연소득
* IE0300007 NICE최종연소득금액_IE0305_000__금융CB

 

    IF A5WC0000008000 > 0 THEN INCOME = A5WC0000008000 ;
    ELSE IF A5WC0000000200 > 0 THEN INCOME = A5WC0000000200 ;
    ELSE IF IE0300007 > 0 THEN INCOME = IE0300007*10 ;
    ELSE INCOME = 0 ;



* 연소득 컬럼 명명
  * INCOME(천원단위)

def _get_year_income(A5WC0000008000, A5WC0000000200, IE0300007):
    year_income = 0
    if A5WC0000008000 > 0:
        # 천원단위
        year_income = A5WC0000008000 ;
    elif A5WC0000000200 > 0:
        # 천원단위
        year_income = A5WC0000000200 ;
    elif IE0300007 > 0:
        # 만원단위여서 * 10
        year_income = IE0300007*10 ;
    else:
        year_income = 0 ;
        
    return year_income

need_cols = ['A5WC0000008000', 'A5WC0000000200', 'IE0300007']

total_df[need_cols].head()

total_df['INCOME'] = total_df[need_cols].apply(lambda x: _get_year_income(x['A5WC0000008000'],
                                                                          x['A5WC0000000200'],
                                                                          x['IE0300007']), axis = 1)

inner_total_df['INCOME'] = total_df['INCOME']

# <font color='orange'>03. 연소득 컬럼 구간 interval 컬럼 생성

```
SEG 1
    0~12000미만 = 1
    12000~16000미만 = 2　 　 
    16000~20000미만 = 3　 
    20000~24000미만 = 4 　 　 
　 　 　 　 　 　 
SEG 2 
    24000~27000미만 = 5
    27000~30000미만 = 6
    30000~33000미만 = 7 　 
    33000~35000미만 = 8
　 　 　 　 
SEG 3 
    35000~40000미만 = 9
    40000~45000미만 = 10
    45000~50000미만 = 11
    50000~55000미만 = 12
    55000~60000미만 = 13
    60000이상 = 14
```

* 연소득 구간화 컬럼 명명
  * INCOME_INTERVAL

### 방법1. pandas cut 사용

income_edge_list= [0,12000,16000,20000,24000,
                   27000,30000,33000,35000,
                  40000,45000,50000,55000,60000,np.inf]

# 구간값 생성
## 구간 경계값 : 최소-이상, 최대-미만
income_interval = pd.cut(total_df['INCOME'], bins = income_edge_list, right=False, )

income_interval

# unique한 구간값 
income_interval_categories = income_interval.cat.categories.values

# 각 대상건의 구간값 코드화
income_interval_codes = income_interval.cat.codes.values
# 각 구간값 1부터 시작하도록 +1
income_interval_codes = income_interval_codes + 1

total_df['INCOME_INTERVAL'] = income_interval_codes

inner_total_df['INCOME_INTERVAL'] = total_df['INCOME_INTERVAL']

### 방법2. 함수 정의

def _get_year_income_interval(year_income):
    year_income_interval = 0
    # SEG 1 
    if 0 <= year_income < 12000:
        year_income_interval = 1
    elif 12000 <= year_income < 16000:
        year_income_interval = 2
    elif 16000 <= year_income < 20000:
        year_income_interval = 3
    elif 20000 <= year_income < 24000:
        year_income_interval = 4
        
    # SEG 2
    elif 24000 <= year_income < 27000:
        year_income_interval = 5
    elif 27000 <= year_income < 30000:
        year_income_interval = 6
    elif 30000 <= year_income < 33000:
        year_income_interval = 7
    elif 33000 <= year_income < 35000:
        year_income_interval = 8
    
    # SEG 3
    elif 35000 <= year_income < 40000:
        year_income_interval = 9
    elif 40000 <= year_income < 45000:
        year_income_interval = 10
    elif 45000 <= year_income < 50000:
        year_income_interval = 11
    elif 50000 <= year_income < 55000:
        year_income_interval = 12
    elif 55000 <= year_income < 60000:
        year_income_interval = 13
    else:
        year_income_interval = 14
        
    return year_income_interval

total_df['INCOME_INTERVAL'] = total_df['INCOME'].apply(lambda x: _get_year_income_interval(x))

inner_total_df['INCOME_INTERVAL'] = total_df['INCOME_INTERVAL']

# <font color='orange'>04. UDIR 구간화 컬럼 생성

* A5RCLSRL013400 웰컴용UDIR 

```
    0~50미만 = 1
    50~100미만 = 2
    100~150미만 = 3
    150~200미만 = 4
    200~250미만 = 5
    250~300이하 = 6
    300초과 = 7
```


* UDIR 구간화 컬럼 명명
  * UDIR_INTERVAL

def _get_udir_interval(A5RCLSRL013400):
    udir_interval = 0
    
    if 0 <= A5RCLSRL013400 < 50:
        udir_interval = 1
    elif 50 <= A5RCLSRL013400 < 100:
        udir_interval = 2
    elif 100 <= A5RCLSRL013400 < 150:
        udir_interval = 3
    elif 150 <= A5RCLSRL013400 < 200:
        udir_interval = 4
    elif 200 <= A5RCLSRL013400 < 250:
        udir_interval = 5
    elif 250 <= A5RCLSRL013400 <= 300:
        udir_interval = 6
    elif 300 < A5RCLSRL013400:
        udir_interval = 7
        
    return udir_interval

total_df['UDIR_INTERVAL'] = total_df['A5RCLSRL013400'].apply(lambda x: _get_udir_interval(x))

inner_total_df['UDIR'] = total_df['A5RCLSRL013400']
inner_total_df['UDIR_INTERVAL'] = total_df['UDIR_INTERVAL']

# <font color='orange'>05. SEG 생성

def _get_seg(year_income_interval, udir_interval):
    seg = 0
    if (year_income_interval in [1,2,3,4]) and (udir_interval in [1,2,3,4]):
        seg = 1
    elif (year_income_interval in [5,6,7,8]) and (udir_interval in [1,2,3,4,5]): 
        seg = 2
    elif (year_income_interval in [9,10,11,12,13,14]) and (udir_interval in [1,2,3,4,5,6]):
        seg = 3
        
    return seg

need_cols = ['INCOME_INTERVAL','UDIR_INTERVAL']

total_df['SEG'] = total_df[need_cols].apply(lambda x: _get_seg(x['INCOME_INTERVAL'],
                                                  x['UDIR_INTERVAL']), axis = 1)

inner_total_df['SEG'] = total_df['SEG']

# <font color='orange'>06. SEG 재구성

* seg1, 2 -> seg1
* seg3 -> 2

def _get_final_seg(seg):
    final_seg = 0
    
    if seg == 1 or seg == 2:
        final_seg = 1
    elif seg == 3:
        final_seg = 2
    return final_seg

total_df['FINAL_SEG'] = total_df['SEG'].apply(lambda x: _get_final_seg(x))

inner_total_df['FINAL_SEG']=total_df['FINAL_SEG'] 

# <font color='orange'>07. 저장

save_path = r'C:\Users\w10\Desktop\신용평가모형 세미나\2주차\데이터\temp'

inner_total_df.to_csv(os.path.join(save_path, 'temp4_inner_total_df.csv'), index = False)
