==> credit evaluation_univariate/code_2nd week/05. CAP_34, 사망자, 개발기간 대상, 최종모형개발대상 선정, .ipynb <==
import pandas as pd
import numpy as np
import os
import warnings
warnings.filterwarnings('ignore')
from datetime import datetime

# <font color='orange'>01. load data

rootpath = r'C:\Users\w10\Desktop\신용평가모형 세미나\2주차\데이터\temp'

os.listdir(rootpath)

total_df = pd.read_pickle(os.path.join(rootpath, 'total_df.pkl'), )

inner_total_df = pd.read_csv(os.path.join(rootpath, 'temp4_inner_total_df.csv'), )

total_df.head(2)

inner_total_df.head(2)

# <font color='orange'>02. CAP_34 선정

* 직장인 최저생계비(12288) 적중 여부

def cap34(job, income):
    result = 0
    if job!='D' and income<=12288:
        result = 1
        
    return result

input_cols = ['직군그룹','INCOME']

inner_total_df['CAP_34'] = inner_total_df[input_cols].apply(lambda x: cap34(x['직군그룹'], x['INCOME']), axis = 1)

# <font color='orange'>03. 사망자 선정

inner_total_df['THE_DEAD'] = total_df['LIVESTAT2']

# <font color='orange'>04. 개발기간 대상 선정

def not_dev_date(ym):
    result = 0
    if ym>=201811 and ym<=201907:
        result = 1
    elif ym>=201911:
        result = 1
        
    return result

inner_total_df['NOT_DEV_DATE_FG'] = inner_total_df['신청월'].apply(lambda x: not_dev_date(x))

# <font color='red'>05. 최종개발대상자 선정

def final_dev_fg(dup, grade_zero, pre_bad, house, seg, cap, dead, date):
    result = 0
    if dup==0 and grade_zero==0 and pre_bad==0 and house==0 and seg!=0 and cap==0 and dead==0 and date==0:
        result = 1
        
    return result

input_cols = ['FINAL_DUP_FG','GRADE_ZERO_FG','TOTAL_PRE_BAD_FG','TOTAL_HOUSE_FG','FINAL_SEG','CAP_34','THE_DEAD','NOT_DEV_DATE_FG']

inner_total_df['DEV_TARGET_FG'] = inner_total_df[input_cols].apply(lambda x: final_dev_fg(x['FINAL_DUP_FG'],
                                                                                         x['GRADE_ZERO_FG'],
                                                                                         x['TOTAL_PRE_BAD_FG'],
                                                                                         x['TOTAL_HOUSE_FG'],
                                                                                         x['FINAL_SEG'],
                                                                                         x['CAP_34'],
                                                                                         x['THE_DEAD'],
                                                                                         x['NOT_DEV_DATE_FG'],),
                                                                   axis = 1)

inner_total_df['DEV_TARGET_FG'].sum()

# <font color='orange'>06. 저장

inner_total_df.to_csv(os.path.join(rootpath, 'temp5_inner_total_df.csv'), index = False)
