==> credit evaluation_univariate/code_2nd week/12. SEG 상세 현황.ipynb <==
import pandas as pd
import numpy as np
import os
import glob
from datetime import datetime
import warnings
warnings.filterwarnings('ignore')

# <font color='orange'>01. load data

data_rootpath = r'C:\Users\w10\Desktop\신용평가모형 세미나\2주차\데이터\dev_data'

os.listdir(data_rootpath)

dev_inner_df = pd.read_csv(os.path.join(data_rootpath, 'model_dev_inner_df.csv'))
valid_inner_df = pd.read_csv(os.path.join(data_rootpath, 'model_valid_inner_df.csv'))

# dictionary에 담기
df_dict = dict()

df_dict['dev_inner'] = dev_inner_df
df_dict['valid_inner'] = valid_inner_df

for key, df in df_dict.items():
    print(key, df.shape)

df_dict['dev_inner'].head(2)

# 모형검증데이터 월별 나누기
df_dict['valid_inner_1908'] = valid_inner_df[valid_inner_df['신청월']==201908]
df_dict['valid_inner_1909'] = valid_inner_df[valid_inner_df['신청월']==201909]
df_dict['valid_inner_1910'] = valid_inner_df[valid_inner_df['신청월']==201910]

## <font color='orange'>01소득구간별 + UDIR 현황

* 주택미보유자 대상으로 해야함

def _pivot_grade(data, grade_var, bad_var, column_var):
    '''
    grade_var = index로 갈 변수
    bad_var = 계산할 변수 
    column_var = column으로 갈 변수
    '''
    piv = data.pivot_table(index = grade_var, columns = column_var, values = [bad_var],
                          aggfunc = ['count','sum'],
                           fill_value = 0,
                          margins = True,
                          margins_name = 'ALL'
                          )
    # 컬럼 정렬
#     piv.sort_index( axis = 1, inplace = True, level = column_var)
    # 컬럼명 변경
    piv.columns = ['TOTAL_CNT_' + column_var + '_' + str(col3) if col1 == 'count' else 'BAD_CNT_' + column_var + '_' + str(col3) for (col1, col2, col3) in piv.columns.tolist()]
    # reset index
    piv.reset_index(inplace = True, )
    
    return piv

seg_udir_dict = dict()
for key, df in df_dict.items():
    df = df[df['TOTAL_PRE_BAD_FG'] == 0]
    this = df[df['TOTAL_HOUSE_FG'] == 0]
    seg_udir_dict[key] = _pivot_grade(this, 'INCOME_INTERVAL','BAD','UDIR_INTERVAL')
    

seg_udir_dict['dev_inner']

save_path = r'C:\Users\w10\Desktop\신용평가모형 세미나\2주차\데이터\SEG상세현황'

# 한 엑셀파일에 여러 sheet로 저장


save_path = os.path.join(save_path, 'SEG상세현황_결과.xlsx')
with pd.ExcelWriter(save_path) as writer:
    for key, result in seg_udir_dict.items():
        result.to_excel(writer, sheet_name = key, )
